Buffered logger uses a throttling buffer
Repeated log messages are logged only a single time with the text ": message
received #{count} times" appended. The default is to buffer and throttle a
single message, so interleaved repeated messages will not be throttled. This
can be changed by setting the auto_flushing value.

This commit changes the default behaviour of the logger from no buffering to
buffering a single message. I believe that this is a desirable change.